+++ Changes.md
@@ -2,10 +2,10 @@
 
 ## TODO
 
-- feat: add and use `pyenv`, `pipenv`, use latest Python+Ansible
 - feat: AWS cli installation/setup
 - docs: add usage descriptions with videos and images in `Usages.md`
 
+- ~feat: add and use `pyenv`, `pipenv`, use latest Python+Ansible~
 - ~feat: support both Ubuntu 20.04 and 22.04 LTS~
 - ~feat: install docker, kubectl, k9s, kind~
 - ~feat(neovim): Install neovim-ruby-host for all rubies~